CLIM-1454: Atom22 Move Window augmentation into the image-rastering namespace
The declare global { interface Window { ... } } block describing $.fn.prepare_raster, mapPrepareRasterPostHttpPayload(Debugger), URL_ENCODER_SALT, and DATA_URL lived in download-map-modal.tsx, an unrelated React component file. map-container.tsx previously type-checked only because that component file happened to be in the same TypeScript program — an accidental dependency invisible to import-based tooling.

Move the block verbatim into apps/src/lib/map/image-rastering/window-globals.d.ts, using import type for Prepare_Raster/PrepareRasterPostHttpPayload (./types) and PrepareRasterPostHttpPayloadDebugger (./create-raster-debugger). The ambient declaration now lives beside the code it describes. Pure structural move, no behaviour change; map-container.tsx untouched.
